Cleaning up

+++ attention_mim.py
@@ -147,7 +147,7 @@ class AttentionMaskingMIM(nn.Module):
             features = self.encoder.encoder(x).last_hidden_state  # [B, N, C]
             attn_map = torch.norm(features, dim=1).view(B, h_0, w_0)  # Approximate attention proxy
 
-        mask = torch.ones_like(x)
+        mask = torch.ones_like(x).long()
         attention_mask = 1.0 - self.generate_attention_mask(attn_map).requires_grad_(True)
         attention_mask = F.interpolate(mask,
                              size=(H, W), mode='nearest').unsqueeze(1)
